chenged all floats to doubles

+++ src/main/java/DBook/project/app/DBookApplication.java
@@ -72,7 +72,7 @@ public class DBookApplication {
 
 		//testowanie Book
 		try(Session s = driver.session(SessionConfig.builder().withDefaultAccessMode(AccessMode.WRITE).build())) {
-			Book b = new Book("Tytus, Romek i Atomek", new Float("29.99"));
+			Book b = new Book("Tytus, Romek i Atomek", new Double("29.99"));
 			s.writeTransaction(b::addToDB);
 		}
